requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.7.0

# Machine Learning
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0

# Deep Learning (Optional)
tensorflow>=2.8.0
torch>=1.12.0

# Time Series Forecasting
prophet>=1.1.0
statsmodels>=0.13.0

# Feature Engineering & Analysis
shap>=0.41.0
feature-engine>=1.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.8.0

# Development & Testing
pytest>=7.0.0
jupyter>=1.0.0
notebook>=6.4.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Optional: For enhanced performance
numba>=0.56.0
joblib>=1.1.0 